Fixed project_tree, now only looking inside src/
If the project root folder contains for example a build/ or a dist/
folder, which may contain some of the sub folders which project_tree is
looking for, then the result may not bee the expected.

This error is fixed by forcing project_tree to search for sub folders
inside the src/ folder only.