Minor bug fixes of original code.
The super call in __init__ of the MaltegoException class was given the
Super class and not the sub class.

The appendelement and removeelement of the _Entity class was trying to
add and remove elements from the underlying fields and labels variables
as if they were lists. However they were actually dictionaries.