Removal of dead code and elimination of a bug.
Dead code:

  self._canari_fields = dict([(f.name, f.value) for f in  self.entity.fields.values()])

We can remove this line of code from the constructor of
MaltegoTransformRequestMessage, as it is basically a no op:

 1. Trying to read from self.entity in __init__ will ALWAYS end
    up returning a 'none-entity' (return Entity('')), as the read
    from self.entities will fail due to the fact that it has not
    been populated with anything yet.

    a. This read from self.entities has the side effect of
       reading from the safedexml List field type, which
       initialises it to the empty list.  When we later on tries
       to actually parse anything into the field, it will fail as
       the call:

          val = super(List, self).__get__(obj)

       inside safedexml.fields.List.parse_child_node will return
       '[]' instead of None.  Because of this, the if statement
       will skip the 'parsing logic' and eventually end up
       returning PARSE_SKIP, instead of PARSE_CHILDREN.

    b. This has the effect that the <entities>...</entities> tag
       is not parsed, and thus the entities List field of the
       request message is newer populated.

 2. '_canari_fields' is not references anywhere else in the canari code.
    So even if the call did return anything, it would newer be used.

 3. To the best of my knowledge the code doesn't produce any side
    effects that are needed or even desired.

With the dead code, the following valid XML from Maltego, will
not populate the entities field.  This is an issue, as the
request message is sent to the transform, from where it is not
possible to access any of the information about the entity (most
importantly the value).
Worse is, if the previous bugfix has been implemented, where the
entities is no longer optional.  Then it will fail with an
ParseError due to the required field not being found.

--------------------------------------------------------

In [1]: import canari.maltego.message as msg

In [2]: mmsg = msg.MaltegoMessage.parse('''
<MaltegoMessage>
   <MaltegoTransformRequestMessage>
      <Entities>
         <Entity Type="Domain">
            <AdditionalFields>
               <Field Name="fqdn" DisplayName="Domain
               Name">paterva.com</Field>
            </AdditionalFields>
            <Weight>0</Weight>
            <Value>paterva.com</Value>
         </Entity>
      </Entities>
      <Limits SoftLimit="255" HardLimit="255"/>
   </MaltegoTransformRequestMessage>
</MaltegoMessage>
''')

--------------------------------------------------------

Now if trying to access the entity property of the request
message, we will get a non-entity back as seen by the empty
value and the empty list of entities.

--------------------------------------------------------

In [3]: mmsg.message.entity.value
Out[3]: ''

In [4]: mmsg.message.entities
Out[4]: []

--------------------------------------------------------

This can be furhter documented by adding debug statements inside
the 'entity' property.  However with the dead code eliminated,
the correct information is returned as seen below.

--------------------------------------------------------

In [3]: mmsg.message.entity.value
Out[3]: u'paterva.com'

In [4]: mmsg.message.entities
Out[4]: [<canari.maltego.message._Entity at 0x7f1ef1b1ff50>]

--------------------------------------------------------

All in all, this seem to be a bug in safedexml.  Currently I have
only noticed it on the List field type, however it may or may not
be present on other field types as well.  This isue is solved by
not reading from the entities field before the user has a chance
to call the parse function or if explicitly adding instances into
it (by appendelement) for use when rendering to xml.